@@ -1,4 +1,4 @@
 import { afterEach } from 'vitest';
-import { enableAutoDestroy } from '@vue/test-utils';
+import { enableAutoUnmount } from '@vue/test-utils';
 
-enableAutoDestroy(afterEach);
+enableAutoUnmount(afterEach);
