@@ -1,9 +1,10 @@
 import { describe, it, expect } from 'vitest';
 import { ISanityCheckOptions } from '@/infrastructure/RuntimeSanity/Common/ISanityCheckOptions';
 import { RuntimeSanityValidator } from '@/presentation/bootstrapping/Modules/RuntimeSanityValidator';
+import { expectDoesNotThrowAsync, expectThrowsAsync } from '@tests/unit/shared/Assertions/ExpectThrowsAsync';
 
 describe('RuntimeSanityValidator', () => {
-  it('calls validator with correct options upon bootstrap', () => {
+  it('calls validator with correct options upon bootstrap', async () => {
     // arrange
     const expectedOptions: ISanityCheckOptions = {
       validateEnvironmentVariables: true,
@@ -15,11 +16,11 @@ describe('RuntimeSanityValidator', () => {
     };
     const sut = new RuntimeSanityValidator(validatorMock);
     // act
-    sut.bootstrap();
+    await sut.bootstrap();
     // assert
     expect(actualOptions).to.deep.equal(expectedOptions);
   });
-  it('propagates the error if validator fails', () => {
+  it('propagates the error if validator fails', async () => {
     // arrange
     const expectedMessage = 'message thrown from validator';
     const validatorMock = () => {
@@ -27,17 +28,17 @@ describe('RuntimeSanityValidator', () => {
     };
     const sut = new RuntimeSanityValidator(validatorMock);
     // act
-    const act = () => sut.bootstrap();
+    const act = async () => { await sut.bootstrap(); };
     // assert
-    expect(act).to.throw(expectedMessage);
+    await expectThrowsAsync(act, expectedMessage);
   });
-  it('runs successfully if validator passes', () => {
+  it('runs successfully if validator passes', async () => {
     // arrange
     const validatorMock = () => { /* NOOP */ };
     const sut = new RuntimeSanityValidator(validatorMock);
     // act
-    const act = () => sut.bootstrap();
+    const act = async () => { await sut.bootstrap(); };
     // assert
-    expect(act).to.not.throw();
+    await expectDoesNotThrowAsync(act);
   });
 });
