@@ -60,11 +60,13 @@ function mountWrapperComponent(scenario?: {
     },
     template: '<div></div>',
   }, {
-    provide: {
-      [InjectionKeys.useCollectionState as symbol]:
-        () => useStateStub.get(),
-      [InjectionKeys.useAutoUnsubscribedEvents as symbol]:
-        () => new UseAutoUnsubscribedEventsStub().get(),
+    global: {
+      provide: {
+        [InjectionKeys.useCollectionState as symbol]:
+          () => useStateStub.get(),
+        [InjectionKeys.useAutoUnsubscribedEvents as symbol]:
+          () => new UseAutoUnsubscribedEventsStub().get(),
+      },
     },
   });
 
